requirements.txt
aiosqlite==0.16.1
asgiref==3.4.1
asyncpg==0.23.0
click==8.0.1
fastapi==0.66.0
h11==0.12.0
iso8601==0.1.14
pydantic==1.8.2
pypika-tortoise==0.1.1
pytz==2021.1
starlette==0.14.2
tortoise-orm==0.17.4
typing-extensions==3.10.0.0
uvicorn==0.14.0
